pyproject.toml
[project]
name = "ai-interview-assistant"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.23.0",
    "beautifulsoup4>=4.14.3",
    "faiss-cpu>=1.9.0",
    "langchain-community>=0.3.0",
    "langchain-core>=1.2.0",
    "langchain-huggingface>=0.1.0",
    "langchain-mistralai>=1.1.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "selenium>=4.39.0",
    "sentence-transformers>=3.0.0",
    "torch>=2.9.1",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "pre-commit>=4.0.1",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.14.9",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
# https://docs.astral.sh/ruff/rules/
select = [
    "E", # pycodestyle (error)
    "F", # Pyflakes
    "I", # isort
    "N", # pep8-naming
    "B", # black
    "W", # pycodestyle (warning)
]
ignore = [
    "N812",  # Example: Lowercase `functions` imported as non-lowercase `F`
    "N803",  # Argument name `X_train` should be lowercase
    "N806",  # Variable `X_train` in function should be lowercase
    "E501",  # Line too long
]

[tool.pytest.ini_options]
markers = [
    "e2e: end-to-end tests that require network and external services (e.g., Mistral API)",
]
